backend/requirements.txt
flask
flask-cors
google-generativeai
gtts
playsound==1.2.2
SpeechRecognition
pyaudio
